Bug 1522675 - use a container with 'native vpc' / 'ip aliases'
Accessing other GCP resources from within a GKE cluster, without using
public IPs, requires that the GKE cluster have native VPC support
enabled.  Basically, that means that the Kubernetes resources operate in
the same networking environment as the rest of GCP and thus traffic can
be routed to other GCP resources.  Such as Cloud SQL DBs.

The key change here is in tf/gke/container-cluster.tf, where the
configuration of ip_allocation_policy forces use of a native VPC.

This parameter requires a newer version of the google provider.

It also seems that there was a bug in the cert-manager dependencies, now
fixed.